Move summary struct into base class for modded class compatibility
