+++ Assets/Scripts/Internal/DaggerfallMobileUnit.cs
@@ -43,9 +43,6 @@ namespace DaggerfallWorkshop
         const int numberOrientations = 8;
         const float anglePerOrientation = 360f / numberOrientations;
 
-        [SerializeField]
-        MobileUnitSummary summary = new MobileUnitSummary();
-
         Camera mainCamera = null;
         MeshFilter meshFilter = null;
         MeshRenderer meshRenderer = null;
@@ -61,11 +58,6 @@ namespace DaggerfallWorkshop
         bool freezeAnims = false;
         bool animReversed = false;
         int frameSpeedDivisor = 1;
-
-        public MobileUnitSummary Summary
-        {
-            get { return summary; }
-        }
 
         public override bool IsSetup
         {
@@ -129,24 +121,6 @@ namespace DaggerfallWorkshop
         {
             get { return Summary.specialTransformationCompleted; }
             protected set { summary.specialTransformationCompleted = value; }
-        }
-
-        [Serializable]
-        public struct MobileUnitSummary
-        {
-            public bool IsSetup;                                        // Flagged true when mobile settings are populated
-            public Rect[] AtlasRects;                                   // Array of rectangles for atlased materials
-            public RecordIndex[] AtlasIndices;                          // Indices into rect array for atlased materials, supports animations
-            public Vector2[] RecordSizes;                               // Size and scale of individual records
-            public int[] RecordFrames;                                  // Number of frames of individual records
-            public MobileEnemy Enemy;                                   // Mobile enemy settings
-            public MobileStates EnemyState;                             // Animation state
-            public MobileAnimation[] StateAnims;                        // Animation frames for this state
-            public MobileBillboardImportedTextures ImportedTextures;    // Textures imported from mods
-            public int AnimStateRecord;                                 // Record number of animation state
-            public int[] StateAnimFrames;                               // Sequence of frames to play for this animation. Used for attacks
-            public byte ClassicSpawnDistanceType;                       // 0 through 6 value read from spawn marker that determines distance at which enemy spawns/despawns in classic.
-            public bool specialTransformationCompleted;                 // Mobile has completed special transformation (e.g. Daedra Seducer)
         }
 
         void Start()
